+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -201,7 +201,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01khb0wkw5pcv9ggyy0var131v"),
+  _id: z.string().default("conv_01khb8dyh17sw23bax20htmzm7"),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),