+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4031,7 +4031,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KHB0WKQ4XPXF7ENHGVEDE7ZF"),
+  id: z.string().default("01KHB8DYCA4K835TVAQF19Y935"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),