+++ packages/orq-rc/RELEASES.md
@@ -6759,3 +6759,13 @@ Based on:
 - [typescript v4.3.0-rc.62] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.62] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.62 - packages/orq-rc
+
+## 2026-02-13 10:26:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.715.0 (2.822.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.63] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.63] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.63 - packages/orq-rc